# fix: Fix growing factor for capacity in Array