--- a/std/assembly/array.ts
+++ b/std/assembly/array.ts
@@ -15,12 +15,13 @@ import { E_INDEXOUTOFRANGE, E_INVALIDLENGTH, E_ILLEGALGENTYPE, E_EMPTYARRAY, E_H
 }
 
 /** Ensures that the given array has _at least_ the specified backing size. */
-function ensureSize(array: usize, minSize: usize, alignLog2: u32): void {
+function ensureCapacity(array: usize, minSize: usize, alignLog2: u32): void {
   // depends on the fact that Arrays mimic ArrayBufferView
   var oldCapacity = changetype<ArrayBufferView>(array).byteLength;
   if (minSize > <usize>oldCapacity >>> alignLog2) {
     if (minSize > BLOCK_MAXSIZE >>> alignLog2) throw new RangeError(E_INVALIDLENGTH);
     let oldData = changetype<usize>(changetype<ArrayBufferView>(array).buffer);
+    // find next power of two size. It usually grows old capacity by factor of two
     let newCapacity = nextPowerOf2(minSize << alignLog2);
     let newData = __renew(oldData, newCapacity);
     memory.fill(newData + oldCapacity, 0, newCapacity - oldCapacity);
@@ -79,7 +80,7 @@ export class Array<T> {
   }
 
   set length(newLength: i32) {
-    ensureSize(changetype<usize>(this), newLength, alignof<T>());
+    ensureCapacity(changetype<usize>(this), newLength, alignof<T>());
     this.length_ = newLength;
   }
 
@@ -115,7 +116,7 @@ export class Array<T> {
   @operator("[]=") private __set(index: i32, value: T): void {
     if (<u32>index >= <u32>this.length_) {
       if (index < 0) throw new RangeError(E_INDEXOUTOFRANGE);
-      ensureSize(changetype<usize>(this), index + 1, alignof<T>());
+      ensureCapacity(changetype<usize>(this), index + 1, alignof<T>());
       this.length_ = index + 1;
     }
     this.__uset(index, value);
@@ -213,8 +214,7 @@ export class Array<T> {
   push(value: T): i32 {
     var length = this.length_;
     var newLength = length + 1;
-    // growFactor == 1 means new capacity will increase by factor of 2
-    ensureSize(changetype<usize>(this), newLength, alignof<T>());
+    ensureCapacity(changetype<usize>(this), newLength, alignof<T>());
     if (isManaged<T>()) {
       store<usize>(this.dataStart + (<usize>length << alignof<T>()), changetype<usize>(value));
       __link(changetype<usize>(this), changetype<usize>(value), true);
@@ -363,7 +363,7 @@ export class Array<T> {
 
   unshift(value: T): i32 {
     var newLength = this.length_ + 1;
-    ensureSize(changetype<usize>(this), newLength, alignof<T>());
+    ensureCapacity(changetype<usize>(this), newLength, alignof<T>());
     var dataStart = this.dataStart;
     memory.copy(
       dataStart + sizeof<T>(),