--- a/std/assembly/array.ts
+++ b/std/assembly/array.ts
@@ -7,7 +7,7 @@ import { idof, isArray as builtin_isArray } from "./builtins";
 import { E_INDEXOUTOFRANGE, E_INVALIDLENGTH, E_ILLEGALGENTYPE, E_EMPTYARRAY, E_HOLEYARRAY } from "./util/error";
 
 // @ts-ignore: decorator
-@inline @lazy const MIN_CAPACITY = 8;
+@inline @lazy const MIN_SIZE = 8;
 
 // NOTE: n should fix in range (1; 1 << 31]
 // @ts-ignore: decorator
@@ -16,13 +16,13 @@ import { E_INDEXOUTOFRANGE, E_INVALIDLENGTH, E_ILLEGALGENTYPE, E_EMPTYARRAY, E_H
 }
 
 /** Ensures that the given array has _at least_ the specified backing size. */
-function ensureCapacity(array: usize, minSize: usize, alignLog2: u32, canGrow: bool = true): void {
+function ensureCapacity(array: usize, mewSize: usize, alignLog2: u32, canGrow: bool = true): void {
   // Depends on the fact that Arrays mimic ArrayBufferView
   var oldCapacity = <usize>changetype<ArrayBufferView>(array).byteLength;
-  if (minSize > oldCapacity >>> alignLog2) {
-    if (minSize > BLOCK_MAXSIZE >>> alignLog2) throw new RangeError(E_INVALIDLENGTH);
+  if (mewSize > oldCapacity >>> alignLog2) {
+    if (mewSize > BLOCK_MAXSIZE >>> alignLog2) throw new RangeError(E_INVALIDLENGTH);
     let oldData = changetype<usize>(changetype<ArrayBufferView>(array).buffer);
-    let newCapacity = max<usize>(minSize, MIN_CAPACITY) << alignLog2;
+    let newCapacity = max<usize>(mewSize, MIN_SIZE) << alignLog2;
     if (canGrow) {
       // Find next power of two size. It usually grows old capacity by factor of two.
       // Make sure we don't reach BLOCK_MAXSIZE for new growed capacity.
@@ -70,8 +70,8 @@ export class Array<T> {
 
   constructor(length: i32 = 0) {
     if (<u32>length > <u32>BLOCK_MAXSIZE >>> alignof<T>()) throw new RangeError(E_INVALIDLENGTH);
-    // reserve capacity for at least MIN_CAPACITY elements
-    var bufferSize = <usize>max(length, MIN_CAPACITY) << alignof<T>();
+    // reserve capacity for at least MIN_SIZE elements
+    var bufferSize = <usize>max(length, MIN_SIZE) << alignof<T>();
     var buffer = changetype<ArrayBuffer>(__new(bufferSize, idof<ArrayBuffer>()));
     memory.fill(changetype<usize>(buffer), 0, bufferSize);
     this.buffer = buffer; // links
@@ -496,7 +496,7 @@ export class Array<T> {
     }
 
     // calculate the byteLength of the resulting backing ArrayBuffer
-    var byteLength = <usize>max(size, MIN_CAPACITY) << usize(alignof<valueof<T>>());
+    var byteLength = <usize>max(size, MIN_SIZE) << usize(alignof<valueof<T>>());
     var outBuffer = changetype<ArrayBuffer>(__new(byteLength, idof<ArrayBuffer>()));
 
     // create the return value and initialize it