… tests/compiler/class.optimized.wat
… tests/compiler/extends-baseaggregate.optimized.wat
… tests/compiler/extends-baseaggregate.untouched.wat
… tests/compiler/infer-array.optimized.wat
… tests/compiler/infer-array.untouched.wat
… tests/compiler/issues/1699.optimized.wat
… tests/compiler/issues/1699.untouched.wat
… tests/compiler/resolve-access.optimized.wat
… tests/compiler/resolve-access.untouched.wat
… tests/compiler/std-wasi/process.optimized.wat
… tests/compiler/std-wasi/process.untouched.wat
… tests/compiler/std/array-access.optimized.wat
… tests/compiler/std/array-access.untouched.wat
… tests/compiler/std/array-literal.optimized.wat
… tests/compiler/std/array-literal.untouched.wat
… tests/compiler/std/array.optimized.wat
… tests/compiler/std/array.ts
… tests/compiler/std/array.untouched.wat
… tests/compiler/std/date.optimized.wat
… tests/compiler/std/date.untouched.wat
… tests/compiler/std/map.optimized.wat
… tests/compiler/std/map.untouched.wat
… tests/compiler/std/set.optimized.wat
… tests/compiler/std/set.untouched.wat
… tests/compiler/std/static-array.optimized.wat
… tests/compiler/std/static-array.untouched.wat
… tests/compiler/std/staticarray.optimized.wat
… tests/compiler/std/staticarray.untouched.wat
… tests/compiler/std/string.optimized.wat
… tests/compiler/std/string.untouched.wat