Fix #1798

- Now internal capacity growing by factor of two.
- Initial capacity changed from `0` to `8` elements

Performance results for sequentially push 1,000,000 items into empty i32 array (`--runtime=incremental`)

|  Browser      |    master<br/> ms    | this PR<br/> ms  |  JavaScript's Array<br/> (baseline), ms  |
| ------------ |----------------:|------------:|------------------------------:|
|  Firefox 88   |   61557             |                  7 |                                              10 |
|  Chrome 90 |   69552            |                  8 |                                            11.4 |

- [x] I've read the contributing guidelines