@dcodeIO 2021-05-01
Two things:
* Skip doubling when setting `.length`, as it doesn't typically imply a series of `.push`es coming afterwards. Memory is precious.
* Account for the case where doubling would overflow maximum block size, and cap to maximum valid block size there. Otherwise growing at half maximum block size would always fail. Probably also needs a check that after all these computations, the new size is actually larger or equal to the originally requested capacity before doubling/capping, so we don't end up with a corrupted object.